requirements.txt
azure-storage-blob
apache-airflow
pyspark
python-dotenv
requests
pandas